JC:
Large reorganisation of the particle system code to be more generic and
faster. In particular, the code now reflects a more correct attitude
towards physics terms. For example, speed and velocity are different
properties of objects and the name confusion of the old code is removed.

- Documented almost everything completely now removing a lot of the
cut and paste work of before.

- Emitters are reorganised to have a common base class of functionality.
All emitters now have a basic speed and variation params. Velocity is not
a property of the base class, but of the derived ones that need it.

- For non-line/point particles, texture coordinates can be provided, along
with a system to change those coordinates over time based on the time.
Allows for the use of a single texture with multiple small images contained
that is used for effects like fire/smoke etc.

- ParticleFunctions now get passed delta T rather than raw time. Makes it
faster for almost all the calculations performed. Physics functions in
particular are now much simpler to compute.